Migrate some resolver functions to StrawberryField
